initial layout changes
