+++ client/packages/common/src/ui/components/loading/InlineSpinner/InlineSpinner.stories.tsx
@@ -3,16 +3,13 @@ import { ComponentStory, ComponentMeta } from '@storybook/react';
 import { InlineSpinner } from './InlineSpinner';
 import { Box } from '@mui/material';
 
-const Template: ComponentStory<typeof InlineSpinner> = ({
-  color,
-  showText,
-}) => (
+const Template: ComponentStory<typeof InlineSpinner> = ({ color }) => (
   <Box>
     <Box style={{ width: 250, height: 100, border: '1px solid green' }}>
       Showing three boxes, all 100 x 250px, with default flex alignment.
     </Box>
     <Box style={{ width: 250, height: 100, border: '1px solid orange' }}>
-      <InlineSpinner color={color} showText={showText} />
+      <InlineSpinner color={color} messageKey="loading" />
     </Box>
     <Box style={{ width: 250, height: 100, border: '1px solid red' }}>
       Center box has the spinner in it showing how the spinner aligns itself
@@ -32,6 +29,6 @@ export default {
 } as ComponentMeta<typeof InlineSpinner>;
 
 Primary.args = { color: 'primary' };
-PrimaryWithText.args = { color: 'primary', showText: true };
+PrimaryWithText.args = { color: 'primary' };
 Secondary.args = { color: 'secondary' };
-SecondaryWithText.args = { color: 'secondary', showText: true };
+SecondaryWithText.args = { color: 'secondary' };
